test: consolidate parser and headless coverage
